feat: Enhance trading strategies and GUI to support fractional trading units
- Updated GUI to reflect changes in trading units from "每手股数" to "交易单位" allowing for fractional shares.
- Modified backend logic to support fractional lot sizes in trading strategies, including mean cost and fixed amount strategies.
- Enhanced stock search functionality to prioritize exact matches and include ETFs and mutual funds.
- Implemented new data normalization methods for consistent OHLCV structure across different data sources.
- Added tests for fractional trading scenarios in both GUI and simulator engine.
- Updated documentation to reflect changes in trading unit handling and data source capabilities.

+++ simulator/simulator.py
@@ -26,7 +26,7 @@ class Simulator:
     - 为实盘交易（real）做好接口准备
     """
 
-    def __init__(self, lot_size: int = 100, init_cash: float = 100000.0, verbose: bool = False):
+    def __init__(self, lot_size: float = 100.0, init_cash: float = 100000.0, verbose: bool = False):
         """初始化模拟器
 
         Args:
@@ -232,7 +232,7 @@ class Simulator:
 def simulate_mean_cost(symbol: str = "600900",
                        start_date: str = "20250101",
                        end_date: Optional[str] = None,
-                       lot_size: int = 100,
+                       lot_size: float = 100.0,
                        init_cash: float = 100000.0,
                        source: str = "auto",
                        progress_callback: Optional[Callable[[int, int], None]] = None,
@@ -254,7 +254,7 @@ def simulate_mean_cost(symbol: str = "600900",
 def simulate_sma(symbol: str = "600900",
                  df=None,
                  period: int = 20,
-                 lot_size: int = 100,
+                 lot_size: float = 100.0,
                  init_cash: float = 100000.0,
                  progress_callback: Optional[Callable[[int, int], None]] = None,
                  trade_price: str = 'open'):
@@ -278,7 +278,7 @@ def simulate_fixed_amount(symbol: str = "600900",
                          start_date: str = "20250101",
                          end_date: Optional[str] = None,
                          fixed_amount: float = 1000.0,
-                         lot_size: int = 100,
+                         lot_size: float = 100.0,
                          init_cash: float = 100000.0,
                          source: str = "auto",
                          verbose: bool = False,
